Build: use printf instead of echo for portability reasons
There are many flavours of echo and it's not guaranteed, for example,
that the escape characters will be handled the same. On my Linux box,
the escape characters are printed as-is, without triggering the intended
color change in the terminal.
printf on the other hand is guaranteed to always interpret the escape
characters.