Components: Infinite list — fix type errors
In React 0.14, refs to built in types such as <div>s are DOM nodes
already, but custom components are not.

findDOMNode doesn't seem to handle built in types well, so we can test
that the ref is probably a DOM node, and use it directly if it is,
otherwise use findDOMNode.